refactor(generator): remove unused code generator package - Deleted the  package and its related  entry-point, as it is no longer required. - Cleaned up  by removing the obsolete  command. - Updated imports across modules to use the  package structure instead of  where applicable.
