+++ cmd/root.go
@@ -3,10 +3,10 @@ package cmd
 import (
 	"context"
 	"fmt"
-	"github.com/rozdolsky33/ocloud/cmd/internal/cmdcreate"
-	"github.com/rozdolsky33/ocloud/cmd/internal/cmdutil"
-	"github.com/rozdolsky33/ocloud/cmd/internal/display"
-	cmdlogger "github.com/rozdolsky33/ocloud/cmd/internal/logger"
+	"github.com/rozdolsky33/ocloud/cmd/shared/cmdcreate"
+	"github.com/rozdolsky33/ocloud/cmd/shared/cmdutil"
+	"github.com/rozdolsky33/ocloud/cmd/shared/display"
+	cmdlogger "github.com/rozdolsky33/ocloud/cmd/shared/logger"
 	"github.com/rozdolsky33/ocloud/internal/app"
 	"github.com/rozdolsky33/ocloud/internal/config/flags"
 	"github.com/spf13/cobra"
