+++ cmd/utils.go
@@ -5,11 +5,9 @@ import (
 )
 
 // isNoContextCommand provides functionality to check if a command doesn't need a full application context
-// This is a simplified version of the previous CommandRegistry, removing unused methods
 func isNoContextCommand() bool {
 	args := os.Args
-	// If no arguments are provided (just the program name), we don't need context
-	// This avoids initialization when just displaying help/usage information
+	// If no arguments are provided, we don't need context
 	if len(args) < 2 {
 		return true
 	}
@@ -37,11 +35,4 @@ func isNoContextCommand() bool {
 	}
 
 	return false
-}
-
-// isRootCommandWithoutSubcommands checks if the command being executed is the root command without any subcommands or flags
-// This is used to determine whether to display the banner and configuration details
-func isRootCommandWithoutSubcommands() bool {
-	args := os.Args
-	return len(args) == 1
 }
